def add_gained
added new function

def add_gained(): it's supposed to be a function that enables the user to be able to track gained money with a save and load system of the user's current balance. Each time there's an expenditure, the current balance will be reduced by the total amount of the expense, and will, of course add to the total current balance if add_gained is used.

also, edited some existing functions' print() statements to better reflect what's happening.

I will push this commit, and take of the rest some time later.