def show_day_total
added new function: def show_day_total

this function takes input of a date and then searches the JSON save file for that specified date. It will then display all the item entries for that date and also display the total expenditure for that date.

It took me 6 hours to figure out how to make the function display the list of items for a specified date. It had to display the corresponding value of the key that the function was calling, which was quite confusing at first because of how JSON objects and arrays needs to be translated first into Python dictionaries and lists. sigh of relief

small note: I would be adding a balance counter soon so that the total balance would be tracked after def add_gained or def add_expense