"balance": current_balance
In this commit, a balance tracker is added.

global variable "current_balance" added for tracking how much money we have left. This will be used for a lot of the functions.

def load_balance
grabs the "balance" key from the JSON file and serializes it into the program. the global variable "current_balance" will be its value

def add_gained():
adds an (integer) amount to the "current_balance".

def check_balance():
checks the global "current_balance" amount

def save_data; now saves "entries" and "balance" into the file. The "entries" dictionary is the one holding all the dates with the list of item entries.

def add_expense: now has an expression that finds the difference between the "current_balance" and "price"s of all the items added.